Initialization engine has been refactored
