+++ examples/example-1.ts
@@ -1,7 +1,7 @@
 import { Controller } from "../mvc-framework/core/decorators/stereotypes/controller/controller.ts";
 import { GET } from "../mvc-framework/core/decorators/stereotypes/controller/routingDecorator.ts";
 import { QueryParam, RouteParam } from "../mvc-framework/core/decorators/stereotypes/controller/parameterDecorator.ts";
-import { MandarineMVC } from "../mvc-framework/mandarineMVC.ts";
+import { MandarineCore } from "../main-core/mandarineCore.ts";
 
 @Controller()
 class TestController {
@@ -32,4 +32,4 @@ export class Boo {
 
 }
 
-new MandarineMVC().run();
+new MandarineCore().MVC().run();
