+++ examples/example-4.ts
@@ -3,10 +3,8 @@ import { Injectable } from "../main-core/dependency-injection/decorators/injecta
 import { Service } from "../main-core/decorators/stereotypes/service/service.ts";
 import { Controller } from "../mvc-framework/core/decorators/stereotypes/controller/controller.ts";
 import { GET } from "../mvc-framework/core/decorators/stereotypes/controller/routingDecorator.ts";
-import { MandarineMVC } from "../mvc-framework/mandarineMVC.ts";
 import { Inject } from "../main-core/dependency-injection/decorators/Inject.ts";
-import { Reflect } from "../main-core/reflectMetadata.ts";
-import { ApplicationContext } from "../main-core/application-context/mandarineApplicationContext.ts";
+import { MandarineCore } from "../mod.ts";
 
 
 export class ManualInjectionService {
@@ -56,4 +54,4 @@ export class Controller1 {
 
 }
 
-new MandarineMVC().run();
+new MandarineCore().MVC().run();
