+++ examples/example-5.ts
@@ -1,7 +1,7 @@
 import { Configuration } from "../main-core/decorators/stereotypes/configuration/configuration.ts";
 import { MandarineStorageHandler } from "../main-core/mandarine-native/sessions/mandarineDefaultSessionStore.ts";
 import { Injectable } from "../main-core/dependency-injection/decorators/injectable.ts";
-import { MandarineMVC } from "../mvc-framework/mandarineMVC.ts";
+import { MandarineCore } from "../mod.ts";
 
 @Configuration()
 export class ConfigClass {
@@ -16,4 +16,4 @@ export class ConfigClass {
 
 }
 
-new MandarineMVC().run();
+new MandarineCore().MVC().run();
